Here's what I've got so far for making tabs
Note that currently all tabs show the same information. I didn't have
time to implement it tonight, but we'd need to set up an OnClick
listener for each new instantiated tab that would reset the title and
info for the inscription to what we want it to be for that particular
tab.