Changes for Notifications and Moving Button
The button to issue a notification is now located on the Search screen
within the menu. Clicking the button issues a notification and sends the
user to the InscriptionActivity, opening to the new tab. I haven't
touched anything to do with the menus. I also just cleaned up some
stuff. Oh, and most of the notification stuff is now in the BaseActivity