Tab-closing, tab-drawing have minor bugs.
Began filling out Inscription activity with elements like DatePicker for Notification Setting and a Checkbox for Found Inscription. Tab-drawing is currently bugged - you have to click on the second tab for it to draw the contents of either. Closing the first tab then the second works, but closing the second before the first produces a crash. Closing all tabs is currently handled by returning to the Base activity.