Working Inscription Text File Saving
The Inscription Tabs now work completely. None are open at the
beginning, but one can add and close test tabs and then switch to other
activities or leave the app, and the same set of inscription tabs with
the same selected one will open when the inscription page is reopened.