Notification Code
The button on the Search page now calls update for the Observer
BackEndConnection. It's still a little buggy -- I haven't yet gotten
clicking the notification to open the app to the Inscription activity,
rather than the last activity open. However, it des seem to add the tab
to the top of the tab list, as we wanted since it's the most recent
inscription, so that's good.