Sound Notifications
There's now an option on the settings page to choose a different
sound/vibrate/silent setting. The result is stored in a
global_preferences file (where we can store the rest of the global
preferences later on, I believe). Each time the SettingsActivity is
created, the spinner is set to the current setting.